requirements.txt
absl-py==1.2.0
asttokens==2.0.8
astunparse==1.6.3
backcall==0.2.0
bsuite==0.3.5
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1663615672595/work/certifi
charset-normalizer==2.1.1
chex==0.1.5
cloudpickle==2.2.0
contextlib2==21.6.0
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
distrax==0.1.2
dm-env==1.5
dm-haiku==0.0.8
dm-launchpad==0.5.2
dm-reverb==0.7.2
dm-tree==0.1.7
etils==0.8.0
executing==1.0.0
flatbuffers==2.0.7
fonttools==4.37.3
frozendict==2.3.4
gast==0.4.0
google-auth==2.11.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
gym==0.26.1
gym-notices==0.0.8
h5py==3.7.0
idna==3.4
imageio==2.22.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipdb==0.13.9
ipython==8.5.0
jax==0.3.17
jaxlib==0.3.15
jedi==0.18.1
jmp==0.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mizani==0.7.4
ml-collections==0.1.1
mock==4.0.3
networkx==2.8.6
numpy==1.22.4
oauthlib==3.2.1
opt-einsum==3.3.0
optax==0.1.3
packaging==21.3
palettable==3.3.0
pandas==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotnine==0.9.0
portpicker==1.5.2
prompt-toolkit==3.0.31
protobuf==3.19.5
# Editable install with no version control (psro==0.0.1)
-e /home/mxsmith/projects/psro
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rlax==0.1.4
rsa==4.9
scikit-image==0.19.3
scipy==1.9.1
six==1.16.0
stack-data==0.5.0
statsmodels==0.13.2
tabulate==0.8.10
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.3
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-probability==0.16.0
termcolor==2.0.1
tifffile==2022.8.12
toml==0.10.2
toolz==0.12.0
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
